Moved data seeding to extension method
